+++ tvision/examples/tutorial/tvguid10.cc
@@ -144,7 +144,7 @@ void TInterior::draw()       // modified for scroller
         if( j < lineCount && lines[j] != 0 )
             {
             char s[maxLineLength];
-            if( delta.x > strlen(lines[j] ) )
+            if( delta.x > (int)strlen(lines[j] ) )
                 s[0] = EOS;
             else
                 {
